tweak lifecycle build imports
